refactor: move components in subfolders
