+++ frontend/src/components/chat/layout.tsx
@@ -2,8 +2,8 @@ import { ReactElement, ReactNode } from 'react';
 import { AuthGuard } from '@/components/auth/auth-guard.tsx';
 import Box from '@mui/material/Box';
 import { ChatProvider } from '@/components/chat/chat-provider.tsx';
-import { ChatView } from './chat-view';
-import { Header } from '@/components/chat/header.tsx';
+import { ChatView } from './view/chat-view';
+import { Header } from './view/header';
 import { Message } from '@/components/chat/types';
 import { dayjs } from '@/lib/dayjs.ts';
 import { usePathname } from '@/hooks/use-pathname.ts';
