+++ frontend/src/components/chat/view/compose.tsx
@@ -1,8 +1,7 @@
 import { ReactElement, useCallback, useContext } from 'react';
-import Box from '@mui/material/Box';
-import { ChatContext } from './chat-context';
+import { ChatContext } from '../chat-context';
 import { MessageAdd } from './message-add';
-import type { MessageType } from './types';
+import type { MessageType } from '../types';
 
 export function ComposeView(): ReactElement {
   const { createMessage } = useContext(ChatContext);
