+++ frontend/src/components/chat/view/header.tsx
